requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
webdriver-manager==4.0.1
undetected-chromedriver==3.5.5
tqdm==4.67.0
xgboost==2.1.4
scikit-learn==1.6.1
tensorflow==2.16.1
matplotlib==3.10.0
python-telegram-bot==22.1
email-service==2.0.0
email-validator==2.2.0
joblib==1.5.0
fastapi==0.115.12
uvicorn==0.34.2
flask==2.2.5
mlflow==2.22.0
plotly==6.1.0
nbformat==5.10.4
psycopg2-binary==2.9.10
SQLAlchemy<2.0.0
aiohttp==3.12.0
lightgbm==4.6.0
catboost==1.2.8